requirements.txt
certifi
pycurl
pyodbc==4.0.26
yagmail==0.11.220
xlrd==1.2.0
PyYAML==5.1.2
